% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{call_vrd_api}
\alias{call_vrd_api}
\title{call_vrd_api()}
\usage{
call_vrd_api(url_, query = NULL, limit = NULL, api_key = NULL)
}
\arguments{
\item{url_}{A string, api url to be hit}

\item{query}{A string, names the resulting object}

\item{limit}{An integer, how many entires to get from the website, default is no limit.}

\item{api_key}{defaults to the one set in the environment, can be pass one manually}
}
\value{
vrd_api class
}
\description{
Helper function for querying API database
}
\examples{
\dontrun{
call_vrd_api()
}
}
