% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_vrd_api}
\alias{clean_vrd_api}
\title{clean_vrd_api()}
\usage{
clean_vrd_api(api_output)
}
\arguments{
\item{api_output}{api response to be cleaned}
}
\value{
dataframe
}
\description{
helper function for cleaning api database
}
\examples{
\dontrun{
clean_vrd_api()
}
}
