% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recall_count.R
\name{count_recall_make}
\alias{count_recall_make}
\title{count_recall_make}
\usage{
count_recall_make(make, manufacturer = FALSE)
}
\arguments{
\item{make}{A string}
}
\value{
integer
}
\description{
count_recall_make
}
\examples{
\dontrun{
count_recall_make('Nissan')
}
}
