% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recall_count.R
\name{count_recall_model}
\alias{count_recall_model}
\title{count_recall_model}
\usage{
count_recall_model(model)
}
\arguments{
\item{model}{A string}
}
\value{
integer
}
\description{
count_recall_model
}
\examples{
\dontrun{
get_recall_model('488')
}
}
