% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecallByNumber.R
\name{get_recall_by_number}
\alias{get_recall_by_number}
\title{get_recall_by_number}
\usage{
get_recall_by_number(recall_number)
}
\arguments{
\item{recall_number}{An integer}
}
\value{
url response
}
\description{
get_recall_by_number
}
\examples{
\dontrun{
get_recall_by_number(1977044)
}
}
