% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recall_data.R
\name{get_recall_make}
\alias{get_recall_make}
\title{get_recall_make}
\usage{
get_recall_make(make, manufacturer = FALSE, limit = 25, partial = FALSE)
}
\arguments{
\item{make}{A string}

\item{limit}{An integer}

\item{partial}{A bool}
}
\value{
dataframe
}
\description{
get_recall_make
}
\examples{
\dontrun{
get_recall_make('Nissan')
}
}
