% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recall_count.R, R/get_recall_data.R
\name{get_recall_model}
\alias{get_recall_model}
\title{get_recall_model}
\usage{
get_recall_model(model, limit = 25, partial = FALSE)

get_recall_model(model, limit = 25, partial = FALSE)
}
\arguments{
\item{model}{A string}

\item{limit}{An integer}

\item{partial}{A bool}
}
\value{
integer

dataframe
}
\description{
get_recall_model

get_recall_model
}
\examples{
\dontrun{
get_recall_model('488')
}
\dontrun{
get_recall_model('488')
}
}
