% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recall_man.R, R/get_recall_num.R
\name{get_recall_num}
\alias{get_recall_num}
\title{get_recall_numr}
\usage{
get_recall_num(recall_number)

get_recall_num(recall_number)
}
\arguments{
\item{recall_number}{An integer}
}
\value{
text output

text output
}
\description{
get_recall_numr

get_recall_num
}
\examples{
\dontrun{
get_recall_num(1977044)
}
\dontrun{
get_recall_num(1977044)
}
}
