% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recall_data.R
\name{get_recall_number}
\alias{get_recall_number}
\title{get_recall_number}
\usage{
get_recall_number(recall_number, limit = 25)
}
\arguments{
\item{recall_number}{An integer}

\item{limit}{An integer}
}
\value{
dataframe
}
\description{
get_recall_number
}
\examples{
\dontrun{
get_recall_number(1977044)
}
}
