% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_data.R
\name{recall_by_make}
\alias{recall_by_make}
\title{recall_by_make}
\usage{
recall_by_make(
  make,
  manufacturer = FALSE,
  start_year = NULL,
  end_year = NULL,
  limit = 25,
  partial = FALSE,
  api_key = NULL
)
}
\arguments{
\item{make}{List string, a list of make names for vehicles}

\item{manufacturer}{A boolean, flag indicating if manufacturer should be used instead of make for the API call}

\item{start_year}{An integer, start of year range, defaults to 1900.}

\item{end_year}{An integer, end of year range, defaults to 2100.}

\item{limit}{An integer, how many entries to get from the website, default is no limit.}

\item{partial}{A boolean, to ask the API for partial matches on the model name.}

\item{api_key}{A string, optional}
}
\value{
A dataframe, collection of responses from the api
}
\description{
Queries the API with a url that it generates from it's parameters.
}
\details{
This function asks the VRD api for recalls for a specific make or manufacturer.
The API can be asked for vehicle names, car seats, tires, etc.
This function can be given a partial name with the partial boolean parameter set to true, and
the year range can be specified. The API key can be passed manually, or if it is not passed
it will use the API key in the R environment, required to be set to VRD_API.
}
\examples{
\dontrun{
recall_by_make('Nissan')
recall_by_make('Maz', end_year=2000, limit=100, partial=TRUE, api_key=API_KEY)
}
}
