% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_data.R
\name{recall_by_model}
\alias{recall_by_model}
\title{recall_by_model}
\usage{
recall_by_model(
  model,
  start_year = NULL,
  end_year = NULL,
  limit = 25,
  partial = FALSE,
  api_key = NULL
)
}
\arguments{
\item{model}{List string, a list of make names for vehicles}

\item{start_year}{An integer, start of year range, defaults to 1900.}

\item{end_year}{An integer, end of year range, defaults to 2100.}

\item{limit}{An integer, how many entries to get from the website, default is no limit.}

\item{partial}{A bool, to determine if we want partial matches on the make parameter}

\item{api_key}{A string, optional}
}
\value{
A dataframe, collection of responses from the api
}
\description{
Queries the API with a url that it generates from it's parameters.
}
\examples{
\dontrun{
recall_by_model('Civic')
recall_by_model('Sub', limit=100, partial=TRUE, api_key=API_KEY)
}
}
