% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_data.R
\name{recall_by_number}
\alias{recall_by_number}
\title{recall_by_number}
\usage{
recall_by_number(recall_number, limit = 25, api_key = NULL)
}
\arguments{
\item{recall_number}{List integer}

\item{limit}{An integer}

\item{api_key}{A string, optional}
}
\value{
dataframe
}
\description{
recall_by_number
}
\examples{
\dontrun{
recall_by_number(1977044)
}
}
