% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_data.R
\name{recall_details}
\alias{recall_details}
\title{Detailed information on recalls searching by recall number(s)}
\usage{
recall_details(recall_number, api_key = NULL)
}
\arguments{
\item{recall_number}{List of recall numbers.}

\item{api_key}{API access key to use, if not set in environment.}
}
\value{
A data.frame of detailed recall information from the Vehicle Recalls
Database. Includes 15 columns.
}
\description{
Returns detailed information on recalls in the Vehicle Recalls Database
based on recall numbers(s).
}
\details{
This function asks the API for details for recall numbers. Recall numbers can be
passed as an integer or a list of integers. Response detail columns are given in
English and in French. The API key can be passed manually, or if it is not passed
it will use the API key in the R environment.
}
\examples{
\dontrun{
recall_details(1977044)
API_KEY <- 'xxxxxxxxxxx'
recall_details(c(2014216, 2013022), api_key = API_KEY)
}
}
