% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_by_manufacturer.R
\name{recall_manufacturer}
\alias{recall_manufacturer}
\title{recall_by_manufacturer}
\usage{
recall_manufacturer(manufacturer, limit = 25, page = 1)
}
\arguments{
\item{manufacturer:}{A car manufacturer name (ie. Honda, Nissan, etc.)
limit: Number of records to return (default is 25)
page: the serach page result to return (default is 1)}
}
\value{
text output
}
\description{
recall_by_manufacturer
}
\examples{
\dontrun{
get_recall_by_number(1977044)
}
}
