% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{call_vrd_api}
\alias{call_vrd_api}
\title{call_vrd_api()}
\usage{
call_vrd_api(url_, query, limit = 25, page = 1)
}
\value{
vrd_api class
}
\description{
helper function for querying api database
}
\examples{
call_vrd_api()
}
