% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_count.R
\name{count_recall_by_make}
\alias{count_recall_by_make}
\title{count_recall_by_make}
\usage{
count_recall_by_make(
  make,
  manufacturer = FALSE,
  start_year = NULL,
  end_year = NULL
)
}
\arguments{
\item{make}{A string}

\item{manufacturer}{bool, with this flag we are asking for the manufacturer not the make}
}
\value{
integer
}
\description{
count_recall_by_make
}
\examples{
\dontrun{
count_recall_by_make('Nissan')
}
}
