% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_count.R
\name{count_recall_by_make}
\alias{count_recall_by_make}
\title{count_recall_by_make}
\usage{
count_recall_by_make(
  make,
  manufacturer = FALSE,
  start_year = NULL,
  end_year = NULL,
  api_key = NULL
)
}
\arguments{
\item{make}{List string, make or manufacturer name(s)}

\item{manufacturer}{A boolean, with this flag we are asking for the manufacturer not the make.}

\item{start_year}{An integer, start of year range, defaults to 1900.}

\item{end_year}{An integer, end of year range, defaults to 2100.}

\item{api_key}{A string, optional}
}
\value{
An integer, the number of recalls for input make in given year range.
}
\description{
Queries the API with a url that it generates from it's parameters.
}
\examples{
\dontrun{
count_recall_by_make('Nissan')
count_recall_by_make('Firestone', manufacturer=TRUE)
count_recall_by_make('Firestone', api_key=API_KEY)
}
}
