% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_count.R
\name{count_recall_by_model}
\alias{count_recall_by_model}
\title{Count of recalls searching by model}
\usage{
count_recall_by_model(
  model,
  start_year = NULL,
  end_year = NULL,
  api_key = NULL
)
}
\arguments{
\item{model}{List of model names.}

\item{start_year}{Start of year range (optional).}

\item{end_year}{End of year range (optional).}

\item{api_key}{API access key to use, if not set in environment.}
}
\value{
A data.frame providing the count of recalls from the Vehicle Recalls.
}
\description{
Returns the count of recalls in the Vehicle Recalls Database based on a
model search term.
}
\details{
Asks the API for the number of recalls for a given model name.
A year range can be specified. The API key can be passed manually, or if it is not passed
it will use the API key in the R environment.
}
\examples{
\dontrun{
count_recall_by_model('civic')
count_recall_by_model(c('Subaru', 'Toyota'), start_year = 2008)
API_KEY <- 'xxxxxxxxxxx'
count_recall_by_model('Sub', end_year = 2000, api_key = API_KEY)
}
}
