% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_count.R
\name{count_recall_by_model}
\alias{count_recall_by_model}
\title{count_recall_by_model}
\usage{
count_recall_by_model(
  model,
  start_year = NULL,
  end_year = NULL,
  api_key = NULL
)
}
\arguments{
\item{model}{List string, make or manufacturer name(s)}

\item{start_year}{An integer, start of year range, defaults to 1900.}

\item{end_year}{An integer, end of year range, defaults to 2100.}

\item{api_key}{A string, optional.}
}
\value{
integer, the number of recalls for input model in given year range.
}
\description{
Queries the API with a url that it generates from it's parameters.
}
\details{
Asks the API for the number of recalls for a given model name.
A year range can be specified. The API key can be passed manually, or if it is not passed
it will use the API key in the R environment.
}
\examples{
\dontrun{
count_recall_by_model('488')
count_recall_by_model('911', start_year=1980, end_year=2005)
}
}
