% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_count.R
\name{count_recall_by_model}
\alias{count_recall_by_model}
\title{count_recall_by_model}
\usage{
count_recall_by_model(model)
}
\arguments{
\item{model}{list string}
}
\value{
integer
}
\description{
count_recall_by_model
}
\examples{
\dontrun{
count_recall_by_model('488')
}
}
