% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_count.R
\name{count_recall_by_years}
\alias{count_recall_by_years}
\title{count_recall_by_years}
\usage{
count_recall_by_years(start_year = 1900, end_year = 2100, api_key = NULL)
}
\arguments{
\item{start_year}{An integer}

\item{end_year}{An integer}

\item{api_key}{A string, optional}
}
\value{
dataframe
}
\description{
count_recall_by_years
}
\examples{
\dontrun{
count_recall_by_years(2010, 2012)
}
}
