% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_count.R
\name{count_recall_by_years}
\alias{count_recall_by_years}
\title{count_recall_by_years}
\usage{
count_recall_by_years(start_year = 1900, end_year = 2100, api_key = NULL)
}
\arguments{
\item{start_year}{An integer, start of year range, defaults to 1900.}

\item{end_year}{An integer, end of year range, defaults to 2100.}

\item{api_key}{A string, optional.}
}
\value{
A dataframe of vehicles for the year range.
}
\description{
Queries the API with a url that it generates from it's parameters.
}
\examples{
\dontrun{
count_recall_by_years(2010, 2012)
count_recall_by_years(2010, 2012, api_key=API_KEY)
}
}
