% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_count.R
\name{count_recall_by_years}
\alias{count_recall_by_years}
\title{Count of recalls searching by year}
\usage{
count_recall_by_years(start_year = 1900, end_year = 2100, api_key = NULL)
}
\arguments{
\item{start_year}{Start of year range (optional).}

\item{end_year}{End of year range (optional).}

\item{api_key}{API access key to use, if not set in environment.}
}
\value{
A data.frame providing the count of recalls from the Vehicle Recalls.
}
\description{
Returns the count of recalls in the Vehicle Recalls Database based on a
year range search term.
}
\details{
Asks the API for the number of recalls for a given year range.
The year range can be specified, or the defaults can be used.
The API key can be passed manually, or if it is not passed
it will use the API key in the R environment.
}
\examples{
\dontrun{
count_recall_by_years(2010, 2012)
API_KEY <- 'xxxxxxxxxxx'
count_recall_by_years(end_year = 1970, api_key = API_KEY)
}
}
