% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recall_man.R
\name{get_recall_manu}
\alias{get_recall_manu}
\title{get_recall_man}
\usage{
get_recall_manu(manufacturer, limit = 25, page = 1)
}
\arguments{
\item{manufacturer}{A string}

\item{limit}{An integer}

\item{page}{An integer}
}
\value{
dataframe
}
\description{
get_recall_man
}
\examples{
\dontrun{
get_recall_manu('Nissan')
}
}
