% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recall_num.R
\name{get_recall_num}
\alias{get_recall_num}
\title{get_recall_num}
\usage{
get_recall_num(recall_number, limit = 25, page = 1)
}
\arguments{
\item{recall_number}{An integer}

\item{limit}{An integer}

\item{page}{An integer}
}
\value{
dataframe
}
\description{
get_recall_num
}
\examples{
\dontrun{
get_recall_num(1977044)
}
}
