% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recall_count.R, R/get_recall_data.R
\name{get_recall_years}
\alias{get_recall_years}
\title{count_recall_years}
\usage{
get_recall_years(start_year = 1900, end_year = 2100, limit = 25)

get_recall_years(start_year = 1900, end_year = 2100, limit = 25)
}
\arguments{
\item{start_year}{An integer}

\item{end_year}{An integer}

\item{limit}{An integer}
}
\value{
dataframe

dataframe
}
\description{
count_recall_years

get_recall_years
}
\examples{
\dontrun{
count_recall_years(2010, 2012)
}
\dontrun{
get_recall_years(2010, 2012)
}
}
