% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_data.R
\name{recall_by_make}
\alias{recall_by_make}
\title{Summary of recalls searching by make or manufacturer}
\usage{
recall_by_make(
  make,
  manufacturer = FALSE,
  start_year = NULL,
  end_year = NULL,
  limit = 25,
  partial = FALSE,
  api_key = NULL
)
}
\arguments{
\item{make}{List of make or manufacturer names. Case insensitive.}

\item{manufacturer}{Logical; if TRUE, manufacturer is searched on instead
of make.}

\item{start_year}{Start of year range (optional).}

\item{end_year}{End of year range (optional).}

\item{limit}{Number indicating how many recall entries should be returned.
Defaults to 25 which is the default of the API.}

\item{partial}{Logical; if TRUE, returns all partial search term matches.}

\item{api_key}{API access key to use, if not set in environment.}
}
\value{
A data.frame of recall summary information from the Vehicle Recalls
Database. Includes six columns.
}
\description{
Returns summary information of recalls in the Vehicle Recalls Database based
on a make or manufacturer search term.
}
\details{
This function asks the VRD API for recalls for a specific make or manufacturer.
The API can be asked for vehicle names, car seats, tires, etc.
This function can be given a partial name with the partial boolean parameter set to true, and
the year range can be specified. The API key can be passed manually, or if it is not passed
it will use the API key in the R environment.
}
\examples{
\dontrun{
recall_by_make('Nissan')
recall_by_make(c('Mazda', 'Toyota'), start_year = 2008, partial = FALSE)
API_KEY <- 'xxxxxxxxxxx'
recall_by_make('Maz', end_year = 2000, limit = 100, api_key = API_KEY)
}
}
