% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_data.R
\name{recall_by_make}
\alias{recall_by_make}
\title{recall_by_make}
\usage{
recall_by_make(make, manufacturer = FALSE, limit = 25, partial = FALSE)
}
\arguments{
\item{make}{A string}

\item{limit}{An integer}

\item{partial}{A bool}
}
\value{
dataframe
}
\description{
recall_by_make
}
\examples{
\dontrun{
recall_by_make('Nissan')
}
}
