% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_data.R
\name{recall_by_make}
\alias{recall_by_make}
\title{recall_by_make}
\usage{
recall_by_make(
  make,
  manufacturer = FALSE,
  start_year = NULL,
  end_year = NULL,
  limit = 25,
  partial = FALSE,
  api_key = NULL
)
}
\arguments{
\item{make}{List string}

\item{manufacturer}{A bool, with this flag we are asking for the manufacturer not the make}

\item{start_year}{An integer}

\item{end_year}{An integer}

\item{limit}{An integer}

\item{partial}{A bool}

\item{api_key}{A string, optional}
}
\value{
dataframe
}
\description{
recall_by_make
}
\examples{
\dontrun{
recall_by_make('Nissan')
}
}
