% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_data.R
\name{recall_by_model}
\alias{recall_by_model}
\title{Summary of recalls searching by model}
\usage{
recall_by_model(
  model,
  start_year = NULL,
  end_year = NULL,
  limit = 25,
  partial = FALSE,
  api_key = NULL
)
}
\arguments{
\item{model}{List of model names.}

\item{start_year}{Start of year range (optional).}

\item{end_year}{End of year range (optional).}

\item{limit}{Number indicating how many recall entries should be returned.
Defaults to 25 which is the default of the API.}

\item{partial}{Logical; if TRUE, returns all partial search term matches.}

\item{api_key}{API access key to use, if not set in environment.}
}
\value{
A data.frame of recall summary information from the Vehicle Recalls
Database. Includes six columns.
}
\description{
Returns summary information of recalls in the Vehicle Recalls Database based
on a model search term.
}
\details{
This function asks the VRD API for recalls for a specific model.
The API can be asked for model names of vehicle , car seats, tires, etc.
This function can be given a partial name with the partial boolean parameter set to true, and
the year range can be specified. The API key can be passed manually, or if it is not passed
it will use the API key in the R environment.
}
\examples{
\dontrun{
recall_by_model('civic')
recall_by_model(c('Subaru', 'Toyota'), start_year = 2008, partial = FALSE)
API_KEY <- 'xxxxxxxxxxx'
recall_by_model('Sub', end_year = 2000, limit = 100, api_key = API_KEY)
}
}
