% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_data.R
\name{recall_by_model}
\alias{recall_by_model}
\title{recall_by_model}
\usage{
recall_by_model(
  model,
  start_year = NULL,
  end_year = NULL,
  limit = 25,
  partial = FALSE
)
}
\arguments{
\item{model}{List string}

\item{start_year}{An integer}

\item{end_year}{An integer}

\item{limit}{An integer}

\item{partial}{A bool}
}
\value{
dataframe
}
\description{
recall_by_model
}
\examples{
\dontrun{
recall_by_model('Civic')
}
}
