% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_data.R
\name{recall_by_number}
\alias{recall_by_number}
\title{recall_by_number}
\usage{
recall_by_number(recall_number, limit = 25, api_key = NULL)
}
\arguments{
\item{recall_number}{List integer, list of recalls numbers to get information on}

\item{limit}{An integer, how many entries to get from the website, default is no limit.}

\item{api_key}{A string, optional}
}
\value{
A dataframe, collection of responses from the api
}
\description{
Queries the API with a url that it generates from it's parameters.
}
\details{
This function asks the VRD api for recalls for list of given recall numbers. It can be given
either a single recall number or a list of recall numbers. The API key can be passed manually, or if it is not passed
it will use the API key in the R environment, required to be set to VRD_API.
}
\examples{
\dontrun{
recall_by_number(1977044)
recall_by_number(c(2014216, 2013022))
}
}
