% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_data.R
\name{recall_by_number}
\alias{recall_by_number}
\title{Summary of recalls searching by recall number(s)}
\usage{
recall_by_number(recall_number, limit = 25, api_key = NULL)
}
\arguments{
\item{recall_number}{List of recall numbers.}

\item{limit}{Number indicating how many recall entries should be returned.
Defaults to 25 which is the default of the API.}

\item{api_key}{API access key to use, if not set in environment.}
}
\value{
A data.frame of recall summary information from the Vehicle Recalls
Database. Includes six columns.
}
\description{
Returns summary information of recalls in the Vehicle Recalls Database based
on recall numbers(s).
}
\details{
This function asks the VRD API for recalls for list of given recall numbers. It can be given
either a single recall number or a list of recall numbers. The API key can be passed manually, or if it is not passed
it will use the API key in the R environment.
}
\examples{
\dontrun{
recall_by_number(1977044)
API_KEY <- 'xxxxxxxxxxx'
recall_by_number(c(2014216, 2013022), api_key = API_KEY)
}
}
