% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_by_number.R
\name{recall_by_number}
\alias{recall_by_number}
\title{recall_by_number}
\usage{
recall_by_number(recall_number)
}
\arguments{
\item{recall_number}{An integer}
}
\value{
text output
}
\description{
recall_by_number
}
\examples{
\dontrun{
get_recall_by_number(1977044)
}
}
