% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_data.R
\name{recall_by_years}
\alias{recall_by_years}
\title{recall_by_years}
\usage{
recall_by_years(start_year = 1900, end_year = 2100, limit = 25)
}
\arguments{
\item{start_year}{An integer}

\item{end_year}{An integer}

\item{limit}{An integer}
}
\value{
dataframe
}
\description{
recall_by_years
}
\examples{
\dontrun{
recall_by_years(2010, 2012)
}
}
