% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_data.R
\name{recall_by_years}
\alias{recall_by_years}
\title{recall_by_years}
\usage{
recall_by_years(start_year = 1900, end_year = 2100, limit = 25, api_key = NULL)
}
\arguments{
\item{start_year}{An integer, start of year range, defaults to 1900.}

\item{end_year}{An integer, end of year range, defaults to 2100.}

\item{limit}{An integer, how many entries to get from the website, default is no limit.}

\item{api_key}{A string, optional}
}
\value{
A dataframe, collection of responses from the API
}
\description{
Queries the API with a url that it generates from it's parameters.
}
\details{
This function asks the VRD API for recalls for a year range.
It will return a list of vehicles, car seats, tires, etc. Can be given a start year
and or a end year. The default for the start and end year are 1900 and 2100.
The API key can be passed manually, or if it is not passed
it will use the API key in the R environment.
}
\examples{
\dontrun{
recall_by_years(2010, 2012, limit=100)
recall_by_years(2010, 2012, api_key=API_KEY)
}
}
