% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_data.R
\name{recall_details}
\alias{recall_details}
\title{recall_details}
\usage{
recall_details(recall_number, limit = 25)
}
\arguments{
\item{recall_number}{List integer}

\item{limit}{An integer}
}
\value{
dataframe
}
\description{
recall_details
}
\examples{
\dontrun{
recall_details(1977044)
}
}
