% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_data.R
\name{recall_details}
\alias{recall_details}
\title{recall_details}
\usage{
recall_details(recall_number, limit = 25, api_key = NULL)
}
\arguments{
\item{recall_number}{List integer}

\item{limit}{An integer, how many entries to get from the website, default is no limit.}

\item{api_key}{A string, optional}
}
\value{
A dataframe, collection of responses from the api
}
\description{
Queries the API with a url that it generates from it's parameters.
}
\examples{
\dontrun{
recall_details(1977044)
recall_details(c(2014216, 2013022))
}
}
